Add purge cache UI to caches page in admin site.

+++ src/net/instaweb/rewriter/rewrite_context_test.cc
@@ -1717,8 +1717,7 @@ TEST_F(RewriteContextTest, CacheExtendCacheableResource) {
                               &headers));
     EXPECT_EQ("a", content);
     EXPECT_STREQ(StringPrintf("max-age=%lld",
-                              static_cast<long long int>(
-                                  ServerContext::kGeneratedMaxAgeMs / 1000)),
+                              ServerContext::kGeneratedMaxAgeMs/1000),
                  headers.Lookup1(HttpAttributes::kCacheControl));
   }
 }
